Seperates classes in files
