+++ Sources/Image.swift
@@ -8,16 +8,6 @@
 
 import Foundation
 
-/// Represents errors that can be thrown within the SwiftGD module.
-///
-/// - invalidImage: Contains the reason this error was thrown.
-/// - invalidColor: Contains the reason this error was thrown.
-public enum Error: Swift.Error {
-    case invalidRasterFormat
-    case invalidImage(reason: String) // The reason this error was thrown
-    case invalidColor(reason: String) // The reason this error was thrown.
-}
-
 // In case you were wondering: it's a class rather than a struct because we need
 // deinit to free the internal GD pointer, and that's only available to classes.
 public class Image {
@@ -301,155 +291,5 @@ extension Image {
     /// - Throws: `Error` if the export of `self` in specified raster format failed.
     public func export(as rasterFormat: ExportableRasterFormat = .png) throws -> Data {
         return try export(using: rasterFormat)
-    }
-}
-
-public struct Point: Equatable {
-	public var x: Int
-	public var y: Int
-
-	public init(x: Int, y: Int) {
-		self.x = x
-		self.y = y
-	}
-
-    public static func == (lhs: Point, rhs: Point) -> Bool {
-        return lhs.x == rhs.x && lhs.y == rhs.y
-    }
-}
-
-public struct Size: Comparable {
-
-	public var width: Int
-	public var height: Int
-
-	public init(width: Int, height: Int) {
-		self.width = width
-		self.height = height
-	}
-
-	public init(width: Int32, height: Int32) {
-		self.width = Int(width)
-		self.height = Int(height)
-	}
-
-    public static func < (lhs: Size, rhs: Size) -> Bool {
-        return lhs.width < rhs.width && lhs.height < rhs.height
-    }
-
-    public static func == (lhs: Size, rhs: Size) -> Bool {
-        return lhs.width == rhs.width && lhs.height == rhs.height
-    }
-}
-
-// MARK: - Color
-
-public class Color {
-
-	public var redComponent: Double
-	public var greenComponent: Double
-	public var blueComponent: Double
-	public var alphaComponent: Double
-
-	public init(red: Double, green: Double, blue: Double, alpha: Double) {
-		redComponent = red
-		greenComponent = green
-		blueComponent = blue
-		alphaComponent = alpha
-	}
-
-	public static let red = Color(red: 1, green: 0, blue: 0, alpha: 1)
-	public static let green = Color(red: 0, green: 1, blue: 0, alpha: 1)
-	public static let blue = Color(red: 0, green: 0, blue: 1, alpha: 1)
-	public static let black = Color(red: 0, green: 0, blue: 0, alpha: 1)
-	public static let white = Color(red: 1, green: 1, blue: 1, alpha: 1)
-}
-
-// MARK: Hexadecimal
-
-extension Color {
-
-    /// The maximum representable integer for each color component.
-    private static let maxHex: Int = 0xff
-
-    /// Initializes a new `Color` instance from a given hexadecimal color string.
-    ///
-    /// Given string will be stripped from a single leading "#", if applicable.
-    /// Resulting string must meet any of the following criteria:
-    ///
-    /// - Is a string with 8-characters and therefore a fully fledged hexadecimal
-    ///   color representation **including** an alpha component. Given value will remain
-    ///   untouched before conversion. Example: `ffeebbaa`
-    /// - Is a string with 6-characters and therefore a fully fledged hexadecimal color
-    ///   representation **excluding** an alpha component. Given RGB color components will
-    ///   remain untouched and an alpha component of `0xff` (opaque) will be extended before
-    ///   conversion. Example: `ffeebb` -> `ffeebbff`
-    /// - Is a string with 4-characters and therefore a shortened hexadecimal color
-    ///   representation **including** an alpha component. Each single character will be
-    ///   doubled before conversion. Example: `feba` -> `ffeebbaa`
-    /// - Is a string with 3-characters and therefore a shortened hexadecimal color
-    ///   representation **excluding** an alpha component. Given RGB color character will
-    ///   be doubled and an alpha of component of `0xff` (opaque) will be extended before
-    ///   conversion. Example: `feb` -> `ffeebbff`
-    ///
-    /// - Parameters:
-    ///   - string: The hexadecimal color string.
-    ///   - leadingAlpha: Indicate whether given string should be treated as ARGB (`true`) or RGBA (`false`)
-    /// - Throws: `.invalidColor` if given string does not match any of the above mentioned critera or is not a valid hex color.
-    public convenience init(hex string: String, leadingAlpha: Bool = false) throws {
-        let string = try Color.sanitize(hex: string, leadingAlpha: leadingAlpha)
-        guard let code = Int(string, radix: 16) else {
-            throw Error.invalidColor(reason: "0x\(string) is not a valid hex color code")
-        }
-        self.init(hex: code, leadingAlpha: leadingAlpha)
-    }
-
-    /// Initializes a new `Color` instance from a given hexadecimal color values.
-    ///
-    /// - Parameters:
-    ///   - color: The hexadecimal color value, incl. red, green, blue and alpha
-    ///   - leadingAlpha: Indicate whether given code should be treated as ARGB (`true`) or RGBA (`false`)
-    public convenience init(hex color: Int, leadingAlpha: Bool = false) {
-        let max = Double(Color.maxHex)
-        let first = Double((color >> 24) & Color.maxHex) / max
-        let second = Double((color >> 16) & Color.maxHex) / max
-        let third = Double((color >>  8) & Color.maxHex) / max
-        let fourth = Double((color >>  0) & Color.maxHex) / max
-        if leadingAlpha {
-            self.init(red: second, green: third, blue: fourth, alpha: first) // ARGB
-        } else {
-            self.init(red: first, green: second, blue: third, alpha: fourth) // RGBA
-        }
-    }
-
-    // MARK: Private helper
-
-    /// Sanitizes given hexadecimal color string (strips # and forms proper length).
-    ///
-    /// - Parameters:
-    ///   - string: The hexadecimal color string to sanitize
-    ///   - leadingAlpha: Indicate whether given and returning string should be treated as ARGB (`true`) or RGBA (`false`)
-    /// - Returns: The sanitized hexadecimal color string
-    /// - Throws: `.invalidColor` if given string is not of proper length
-    private static func sanitize(hex string: String, leadingAlpha: Bool) throws -> String {
-
-        // Drop leading "#" if applicable
-        var string = string.hasPrefix("#") ? String(string.dropFirst(1)) : string
-
-        // Evaluate if short code w/wo alpha (e.g. `feb` or `feb4`). Double up the characters if so.
-        if string.count == 3 || string.count == 4 {
-            string = string.map({ "\($0)\($0)" }).joined()
-        }
-
-        // Evaluate if this is a fully fledged code with or without alpha (e.g. `ffaabb` or `ffaabb44`), otherwise throw error
-        switch string.count {
-        case 6: // Hex color code without alpha (e.g. ffeeaa)
-            let alpha = String(Color.maxHex, radix: 16) // 0xff (opaque)
-            return leadingAlpha ? alpha + string : string + alpha
-        case 8: // Fully fledged hex color including alpha (e.g. eebbaa44)
-            return string
-        default:
-            throw Error.invalidColor(reason: "0x\(string) has invalid hex color string length")
-        }
     }
 }
